fix intlong

+++ DTOs/Notification/NotificationDTO.cs
@@ -5,7 +5,7 @@ public class NotificationDTO : BaseModelDTO
     public int EmployeeId { get; set; }
     public string EmployeeName { get; set; } = null!;
 
-    public int NotificationCategoryId { get; set; }
+    public long NotificationCategoryId { get; set; }
     public string NotificationCategoryName { get; set; } = null!;
 
     public string Title { get; set; } = null!;
@@ -20,7 +20,7 @@ public class NotificationCreateDTO : BaseModelCreateDTO
 {
     public int EmployeeId { get; set; }
 
-    public int NotificationCategoryId { get; set; }
+    public long NotificationCategoryId { get; set; }
     public string Title { get; set; } = string.Empty;
     public string Message { get; set; } = string.Empty;
     public string? Url { get; set; }
@@ -29,7 +29,7 @@ public class NotificationCreateDTO : BaseModelCreateDTO
 
 public class NotificationUpdateDTO : BaseModelUpdateDTO
 {
-    public int NotificationCategoryId { get; set; }
+    public long NotificationCategoryId { get; set; }
     public string Title { get; set; } = string.Empty;
     public string Message { get; set; } = string.Empty;
     public string? Url { get; set; }
